requirements.txt
tensorflow==2.15.0
dm-reverb==0.14.0
tf-agents==0.19.0
pygame==2.1.3
imageio==2.4.0
pyvirtualdisplay
pyglet
gym[box2d]
pyopengl
pybullet
gym[mujoco]
